Refactor Shamir's Secret Sharing implementation
Modularized the implementation by separating responsibilities into `ShamirsSecretSharing` and `SecretReconstructor` classes.
Introduced `ISecurityLevelManager` for managing security level logic.
Simplified constructors and removed redundant code to improve maintainability and readability.

Resolves: No entry
