+++ src/Cryptography/ShamirsSecretSharing/SecretReconstructor`1.cs
@@ -1,0 +1,47 @@
+// ----------------------------------------------------------------------------
+// <copyright file="ShamirsSecretSharing`1.cs" company="Private">
+// Copyright (c) 2025 All Rights Reserved
+// </copyright>
+// <author>Sebastian Walther</author>
+// <date>10/03/2025 01:42:17 AM</date>
+// ----------------------------------------------------------------------------
+
+#region License
+// ----------------------------------------------------------------------------
+// Copyright 2025 Sebastian Walther
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+#endregion
+
+namespace SecretSharingDotNet.Cryptography.ShamirsSecretSharing;
+
+using Math;
+
+/// <summary>
+/// Represents a class that reconstructs a secret based on the principles of
+/// Shamir's Secret Sharing scheme.
+/// </summary>
+/// <typeparam name="TNumber">
+/// The mathematical type used in the secret reconstruction process.
+/// </typeparam>
+public class SecretReconstructor<TNumber> : SecretReconstructor<TNumber, IExtendedGcdAlgorithm<TNumber>, ExtendedGcdResult<TNumber>>
+{
+    /// <inheritdoc />
+    public SecretReconstructor(IExtendedGcdAlgorithm<TNumber> extendedGcd) : base(extendedGcd) { }
+}
