+++ src/Resources/ErrorMessages.resx
@@ -57,4 +57,7 @@
     <data name="NoMaximumY" xml:space="preserve">
         <value>Unexpected error. The largest Y coordinate could not be determined. Please check the shares.</value>
     </data>
+    <data name="NoValidMersennePrimeExponentAvailable" xml:space="preserve">
+        <value>No valid Mersenne prime exponent &gt;= {0} available.</value>
+    </data>
 </root>
