+++ tests/Cryptography/SharesTest.cs
@@ -1,0 +1,120 @@
+﻿// ----------------------------------------------------------------------------
+// <copyright file="SharesTest.cs" company="Private">
+// Copyright (c) 2019 All Rights Reserved
+// </copyright>
+// <author>Sebastian Walther</author>
+// <date>04/20/2019 10:52:28 PM</date>
+// ----------------------------------------------------------------------------
+
+#region License
+// ----------------------------------------------------------------------------
+// Copyright 2019 Sebastian Walther
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+#endregion
+
+namespace SecretSharingDotNetTest.Cryptography;
+
+using SecretSharingDotNet.Cryptography;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Numerics;
+using Xunit;
+
+/// <summary>
+/// Unit test of the <see cref="Shares{TNumber}"/> class.
+/// </summary>
+public class SharesTest
+{
+    /// <summary>
+    /// Tests the cast from <see cref="string"/> array to <see cref="Shares{TNumber}"/> and vice versa.
+    /// </summary>
+    [Fact]
+    public void TestSharesToStringArray()
+    {
+        // Arrange
+        var stringArray = TestData.GetPredefinedShares();
+
+        // Act
+        Shares<BigInteger> shares = stringArray;
+
+        // Assert
+        Assert.Equal(TestData.GetPredefinedShares(), (string[])shares);
+    }
+
+    /// <summary>
+    /// Tests the cast from <see cref="string"/> to <see cref="Shares{TNumber}"/> and vice versa.
+    /// </summary>
+    [Fact]
+    public void TestSharesToString()
+    {
+        // Arrange
+        var text = string.Join(Environment.NewLine, TestData.GetPredefinedShares()) + Environment.NewLine;
+
+        // Act
+        Shares<BigInteger> shares = text;
+
+        // Assert
+        Assert.Equal(text, shares.ToString());
+    }
+
+    /// <summary>
+    /// Tests the Contains method of the <see cref="Shares{TNumber}"/> class.
+    /// </summary>
+    [Fact]
+    public void TestShareContains()
+    {
+        // Arrange
+        var stringArray = TestData.GetPredefinedShares();
+
+        // Act
+        Shares<BigInteger> shares = stringArray;
+
+        // Assert
+        Assert.Contains(new FinitePoint<BigInteger>(TestData.GetPredefinedShares()[0]), shares);
+    }
+
+    /// <summary>
+    /// Tests the <see cref="IEnumerator{T}"/> implementation of the <see cref="SharesEnumerator{TNumber}"/> class in the <see cref="Shares{TNumber}"/> class.
+    /// </summary>
+    [Fact]
+    public void TestSharesEnumerator()
+    {
+        // Arrange
+        Shares<BigInteger> shares = TestData.GetPredefinedShares();
+        var testDataSequence = TestData.GetPredefinedShares().Select(entry => new FinitePoint<BigInteger>(entry));
+        var testDataArray = testDataSequence as FinitePoint<BigInteger>[] ?? testDataSequence.ToArray();
+
+        // Act
+        var actual = ((IEnumerable)shares).GetEnumerator();
+        using var disposable = actual as IDisposable;
+        var expected = testDataArray.GetEnumerator();
+
+        // Assert
+        for (var i = 0; i < testDataArray.Length; i++)
+        {
+            Assert.Equal(expected.MoveNext(), actual.MoveNext());
+            Assert.Equal(expected.Current, actual.Current);
+        }
+
+        Assert.True(shares.SequenceEqual(testDataArray));
+    }
+}
