+++ tests/SecretSharingDotNetTest.csproj
@@ -24,6 +24,7 @@
     <ProjectReference Include="..\src\SecretSharingDotNet.csproj" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="Moq" Version="4.20.72" />
     <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.3">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
